requirements.txt
streamlit==1.42.0
pandas==2.2.2
python-dotenv==1.0.1
groq==0.18.0
semantic-router==0.1.1
chromadb==0.6.3
sentence_transformers==3.4.1
pysqlite3-binary